Collapse band-name-generator word lists into one data module
The 26 one-array-per-file word lists under band-name-generator/ are now
a single word-pools.ts. Pool names and the public re-exports from
band-name-generator.ts are unchanged.
