+++ src/shared/band-name-generator.ts
@@ -2,42 +2,13 @@
  * Procedural rock/heavy-metal band-name, venue, and gig-description
  * generator for sample data.
  *
- * Word pools live as one-file-per-list under `band-name-generator/`. This
+ * Word pools live in `band-name-generator/word-pools.ts`. This
  * module composes them with a small set of slot-heavy templates and a
  * seeded PRNG (mulberry32) so every output is deterministic given a seed
  * yet the variety is enormous.
  */
 
-import { BAND_ADJECTIVES } from "#shared/band-name-generator/adjectives.ts";
-import { AGE_NOTES } from "#shared/band-name-generator/age-notes.ts";
-import { AUDIENCE_OUTCOMES } from "#shared/band-name-generator/audience-outcomes.ts";
-import { BAND_DESCRIPTORS } from "#shared/band-name-generator/band-descriptors.ts";
-import { BAND_VERBS } from "#shared/band-name-generator/band-verbs.ts";
-import { BUILDING_STATES } from "#shared/band-name-generator/building-states.ts";
-import { CONNECTORS } from "#shared/band-name-generator/connectors.ts";
-import { CROSSOVERS } from "#shared/band-name-generator/crossovers.ts";
-import { EVENT_TYPES } from "#shared/band-name-generator/event-types.ts";
-import { FESTIVAL_TYPES } from "#shared/band-name-generator/festival-types.ts";
-import { GENRES } from "#shared/band-name-generator/genres.ts";
-import { INTENSITIES } from "#shared/band-name-generator/intensities.ts";
-import { NOISE_VERBS } from "#shared/band-name-generator/noise-verbs.ts";
-import { BAND_NOUNS } from "#shared/band-name-generator/nouns.ts";
-import { NUMBER_WORDS } from "#shared/band-name-generator/number-words.ts";
-import { ODD_INSTRUMENTS } from "#shared/band-name-generator/odd-instruments.ts";
-import { BAND_PERSON_NAMES } from "#shared/band-name-generator/person-names.ts";
-import { PRIZES } from "#shared/band-name-generator/prizes.ts";
-import { PROHIBITIONS } from "#shared/band-name-generator/prohibitions.ts";
-import { REQUIREMENTS } from "#shared/band-name-generator/requirements.ts";
-import { SHOW_ITEMS } from "#shared/band-name-generator/show-items.ts";
-import { BAND_SUFFIXES } from "#shared/band-name-generator/suffixes.ts";
-import { TIMES } from "#shared/band-name-generator/times.ts";
-import { TOUR_ADJECTIVES } from "#shared/band-name-generator/tour-adjectives.ts";
-import { VENUE_TYPES } from "#shared/band-name-generator/venue-types.ts";
-import { WEIRD_VENUES } from "#shared/band-name-generator/weird-venues.ts";
-
-// Re-export every pool so existing call sites and tests can keep using
-// `import { BAND_ADJECTIVES } from "#shared/band-name-generator.ts"`.
-export {
+import {
   AGE_NOTES,
   AUDIENCE_OUTCOMES,
   BAND_ADJECTIVES,
@@ -64,7 +35,11 @@ export {
   TOUR_ADJECTIVES,
   VENUE_TYPES,
   WEIRD_VENUES,
-};
+} from "#shared/band-name-generator/word-pools.ts";
+
+// Re-export every pool so existing call sites and tests can keep using
+// `import { BAND_ADJECTIVES } from "#shared/band-name-generator.ts"`.
+export * from "#shared/band-name-generator/word-pools.ts";
 
 /** A pseudo-random function returning a value in [0, 1) */
 export type Rand = () => number;
@@ -138,8 +113,9 @@ const fixArticles = (s: string): string =>
  * that begin with a lowercase pool entry still read as a proper sentence.
  */
 const renderTemplate = (template: string, rand: Rand): string => {
-  const filled = template.replace(/\{(\w+)\}/g, (_, key: string) =>
-    pickFrom(rand, SLOT_POOLS[key]!),
+  const filled = template.replace(
+    /\{(\w+)\}/g,
+    (_, key: string) => pickFrom(rand, SLOT_POOLS[key]!),
   );
   const articled = fixArticles(filled);
   return articled.charAt(0).toUpperCase() + articled.slice(1);
@@ -231,8 +207,7 @@ export const DEFAULT_VENUE_SEED = 0x57a6ed00;
 export const DEFAULT_DESCRIPTION_SEED = 0xde5c0f6c;
 
 const seededList =
-  (gen: (rand: Rand) => string) =>
-  (count: number, seed: number): string[] => {
+  (gen: (rand: Rand) => string) => (count: number, seed: number): string[] => {
     const rand = createRand(seed);
     return Array.from({ length: count }, () => gen(rand));
   };
