--- src/shared/band-name-generator/adjectives.ts
@@ -1,221 +1,0 @@
-/** Weird, evocative adjectives suitable for metal/rock band names */
-export const BAND_ADJECTIVES = [
-  "Brutal",
-  "Equine",
-  "Mystic",
-  "Screaming",
-  "Savage",
-  "Eternal",
-  "Forbidden",
-  "Cursed",
-  "Crystal",
-  "Velvet",
-  "Iron",
-  "Black",
-  "Crimson",
-  "Ghastly",
-  "Howling",
-  "Roaring",
-  "Whispering",
-  "Burning",
-  "Frozen",
-  "Molten",
-  "Electric",
-  "Toxic",
-  "Radioactive",
-  "Phantom",
-  "Spectral",
-  "Skeletal",
-  "Demonic",
-  "Angelic",
-  "Holy",
-  "Unholy",
-  "Profane",
-  "Sacred",
-  "Vengeful",
-  "Wretched",
-  "Tragic",
-  "Tortured",
-  "Twisted",
-  "Bleeding",
-  "Pale",
-  "Withered",
-  "Reanimated",
-  "Drunken",
-  "Feral",
-  "Rabid",
-  "Mighty",
-  "Tyrannical",
-  "Imperial",
-  "Royal",
-  "Divine",
-  "Apocalyptic",
-  "Cosmic",
-  "Lunar",
-  "Solar",
-  "Stellar",
-  "Astral",
-  "Voidborn",
-  "Infernal",
-  "Diabolical",
-  "Sinister",
-  "Cataclysmic",
-  "Volcanic",
-  "Glacial",
-  "Granite",
-  "Marbled",
-  "Gilded",
-  "Tarnished",
-  "Mercurial",
-  "Mongolian",
-  "Jurassic",
-  "Chrome",
-  "Diesel",
-  "Atomic",
-  "Neon",
-  "Plastic",
-  "Wooden",
-  "Rusted",
-  "Festering",
-  "Slithering",
-  "Glowing",
-  "Magnetic",
-  "Thunderous",
-  "Subterranean",
-  "Hyperborean",
-  "Perpetual",
-  "Hooded",
-  "Velour",
-  "Magic",
-  "Naked",
-  "Hidden",
-  "Drowning",
-  "Choking",
-  "Greasy",
-  "Dusty",
-  "Rotting",
-  "Yelping",
-  "Polite",
-  "Ancient",
-  "Vintage",
-  "Final",
-  "Last",
-  "Crawling",
-  "Lurking",
-  "Hovering",
-  "Disgraced",
-  "Disowned",
-  "Defeated",
-  "Vanquished",
-  "Resurrected",
-  "Banished",
-  "Exiled",
-  "Wandering",
-  "Lost",
-  "Hunted",
-  "Haunted",
-  "Possessed",
-  "Anointed",
-  "Forsworn",
-  "Forgotten",
-  "Buried",
-  "Entombed",
-  "Enchanted",
-  "Bewitched",
-  "Spellbound",
-  "Hexed",
-  "Doomed",
-  "Damned",
-  "Glorious",
-  "Sublime",
-  "Foul",
-  "Putrid",
-  "Rancid",
-  "Acrid",
-  "Smouldering",
-  "Sizzling",
-  "Roasting",
-  "Boiling",
-  "Stewing",
-  "Bubbling",
-  "Sputtering",
-  "Crackling",
-  "Flickering",
-  "Faltering",
-  "Trembling",
-  "Shuddering",
-  "Quivering",
-  "Quaking",
-  "Wailing",
-  "Snarling",
-  "Gnashing",
-  "Spitting",
-  "Hissing",
-  "Buzzing",
-  "Droning",
-  "Squawking",
-  "Squelching",
-  "Gurgling",
-  "Slurping",
-  "Sticky",
-  "Knotted",
-  "Tangled",
-  "Crooked",
-  "Crumpled",
-  "Crumbling",
-  "Shattered",
-  "Splintered",
-  "Cracked",
-  "Fractured",
-  "Ruined",
-  "Decrepit",
-  "Derelict",
-  "Forlorn",
-  "Bereft",
-  "Abandoned",
-  "Forsaken",
-  "Regrettable",
-  "Ill-Advised",
-  "Untrustworthy",
-  "Suspicious",
-  "Unhinged",
-  "Bewildered",
-  "Befuddled",
-  "Flummoxed",
-  "Ungovernable",
-  "Inconsolable",
-  "Unbothered",
-  "Begrudging",
-  "Intolerable",
-  "Embarrassing",
-  "Implausible",
-  "Improbable",
-  "Unspeakable",
-  "Inexplicable",
-  "Unrepentant",
-  "Belligerent",
-  "Insatiable",
-  "Inevitable",
-  "Indomitable",
-  "Insufferable",
-  "Impervious",
-  "Imperishable",
-  "Incandescent",
-  "Indignant",
-  "Truculent",
-  "Petulant",
-  "Diminutive",
-  "Gargantuan",
-  "Colossal",
-  "Lopsided",
-  "Bedraggled",
-  "Bewhiskered",
-  "Bespectacled",
-  "Cantankerous",
-  "Pernicious",
-  "Lugubrious",
-  "Saccharine",
-  "Melodious",
-  "Pugilistic",
-  "Sclerotic",
-] as const;
