--- src/shared/band-name-generator/age-notes.ts
@@ -1,18 +1,0 @@
-/** Audience-eligibility blurbs */
-export const AGE_NOTES = [
-  "all ages welcome",
-  "18+ only",
-  "14+ with parent",
-  "family-friendly",
-  "parental advisory in effect",
-  "strictly grown-ups",
-  "over-30s strongly encouraged",
-  "no under-fives without ear defenders",
-  "pets on leads only",
-  "ID checks at the door",
-  "members and accompanied guests",
-  "carer admitted free",
-  "no babes in arms after 9pm",
-  "well-behaved teens permitted",
-  "responsible adults preferred",
-] as const;
