--- src/shared/band-name-generator/audience-outcomes.ts
@@ -1,27 +1,0 @@
-/** Predictable consequences of attending the gig */
-export const AUDIENCE_OUTCOMES = [
-  "regret",
-  "glory",
-  "tinnitus",
-  "sweat",
-  "tears",
-  "mild concern",
-  "confusion",
-  "broken eardrums",
-  "lost dignity",
-  "hoarse voices",
-  "lost car keys",
-  "aching necks",
-  "mismatched footwear",
-  "sticky floors",
-  "questionable bruises",
-  "slightly damp denim",
-  "sudden tinnitus",
-  "minor brain fog",
-  "unexplained leaves in pockets",
-  "inappropriate enthusiasm",
-  "newfound dietary requirements",
-  "missed last buses",
-  "incoherent stories",
-  "second-hand glitter",
-] as const;
