--- src/shared/band-name-generator/band-descriptors.ts
@@ -1,33 +1,0 @@
-/** Collective nouns for "the band's reputation" */
-export const BAND_DESCRIPTORS = [
-  "pioneers",
-  "warriors",
-  "legends",
-  "veterans",
-  "prophets",
-  "godfathers",
-  "stalwarts",
-  "diehards",
-  "originators",
-  "cult heroes",
-  "mavericks",
-  "renegades",
-  "outlaws",
-  "old hands",
-  "elder statesmen",
-  "scene leaders",
-  "underdogs",
-  "scene-stealers",
-  "deep-cut specialists",
-  "self-styled visionaries",
-  "improbable survivors",
-  "eternal underdogs",
-  "long-suffering road dogs",
-  "near-mythical figures",
-  "barely-functional collective",
-  "loosely-affiliated crew",
-  "occasionally-touring side project",
-  "perennially-overlooked outfit",
-  "criminally-undercelebrated trio",
-  "begrudgingly-respected veterans",
-] as const;
