--- src/shared/band-name-generator/band-verbs.ts
@@ -1,27 +1,0 @@
-/** Verbs the band itself does — used in "{bandDescriptor} {bandVerb} ..." patterns */
-export const BAND_VERBS = [
-  "reunite",
-  "return",
-  "regroup",
-  "reform",
-  "ascend",
-  "descend",
-  "emerge",
-  "materialise",
-  "awaken",
-  "surface",
-  "converge",
-  "rally",
-  "assemble",
-  "manifest",
-  "reappear",
-  "convene",
-  "reconvene",
-  "stagger back",
-  "shamble forth",
-  "limp onstage",
-  "march in",
-  "stomp out",
-  "drift in",
-  "crash through the door",
-] as const;
