--- src/shared/band-name-generator/building-states.ts
@@ -1,26 +1,0 @@
-/** Adjective phrases describing the state of a "weird venue" before being repurposed */
-export const BUILDING_STATES = [
-  "converted",
-  "repurposed",
-  "abandoned",
-  "decommissioned",
-  "gentrified",
-  "restored",
-  "haunted",
-  "condemned",
-  "derelict",
-  "half-finished",
-  "recently-evacuated",
-  "formerly-respectable",
-  "barely-legal",
-  "improbably-listed",
-  "informally-occupied",
-  "structurally-unsound",
-  "reluctantly-available",
-  "council-owned",
-  "privately-disowned",
-  "long-disused",
-  "freshly-tagged",
-  "improvisationally-rewired",
-  "inadvisably-reopened",
-] as const;
