--- src/shared/band-name-generator/crossovers.ts
@@ -1,37 +1,0 @@
-/** Themes for crossover/novelty acts */
-export const CROSSOVERS = [
-  "sea-shanty",
-  "druid",
-  "viking",
-  "pirate",
-  "robot",
-  "ghost",
-  "morris-dance",
-  "bagpipe",
-  "yodelling",
-  "yacht-rock",
-  "monster-truck",
-  "sci-fi",
-  "wild-west",
-  "Tudor-court",
-  "Edwardian-music-hall",
-  "Roman-legion",
-  "samurai",
-  "lumberjack",
-  "tea-party",
-  "coal-miner",
-  "deep-sea-trawler",
-  "haunted-circus",
-  "post-apocalyptic",
-  "Cold-War",
-  "Steampunk",
-  "Cyberpunk",
-  "renaissance-fair",
-  "1970s-prog",
-  "1980s-disco",
-  "1990s-rave",
-  "early-noughties-emo",
-  "small-claims-court",
-  "town-council",
-  "garden-centre",
-] as const;
