--- src/shared/band-name-generator/event-types.ts
@@ -1,30 +1,0 @@
-/** What the event itself is — sits where words like "show" / "matinee" used to be hard-coded */
-export const EVENT_TYPES = [
-  "night",
-  "matinee",
-  "show",
-  "evening",
-  "gig",
-  "set",
-  "recital",
-  "séance",
-  "summoning",
-  "ritual",
-  "vigil",
-  "broadcast",
-  "takeover",
-  "residency",
-  "occupation",
-  "livestream",
-  "riot",
-  "jamboree",
-  "knees-up",
-  "happening",
-  "spectacle",
-  "exhibition",
-  "convocation",
-  "wake",
-  "send-off",
-  "reckoning",
-  "exorcism",
-] as const;
