--- src/shared/band-name-generator/festival-types.ts
@@ -1,45 +1,0 @@
-/** Outdoor / event-style venue archetypes */
-export const FESTIVAL_TYPES = [
-  "Festival",
-  "Fest",
-  "Fair",
-  "Carnival",
-  "Convention",
-  "Summit",
-  "Showcase",
-  "Massacre",
-  "Slaughter",
-  "Bash",
-  "Jamboree",
-  "Riot",
-  "Reckoning",
-  "Pilgrimage",
-  "Procession",
-  "Reunion",
-  "Tour",
-  "Crusade",
-  "Onslaught",
-  "Weekender",
-  "All-Dayer",
-  "Knees-Up",
-  "Hoedown",
-  "Stomp",
-  "Throwdown",
-  "Free-For-All",
-  "Marathon",
-  "Wakefield",
-  "Vigil",
-  "Spectacle",
-  "Extravaganza",
-  "Gala",
-  "Soirée",
-  "Roadshow",
-  "Symposium",
-  "Conclave",
-  "Assembly",
-  "Gathering",
-  "Reunion Tour",
-  "Exhibition",
-  "Tournament",
-  "Olympiad",
-] as const;
