--- src/shared/band-name-generator/genres.ts
@@ -1,66 +1,0 @@
-/** Sub-genres used inside gig-description templates */
-export const GENRES = [
-  "doom",
-  "sludge",
-  "prog",
-  "hardcore",
-  "glam",
-  "thrash",
-  "death-metal",
-  "black-metal",
-  "stoner-rock",
-  "folk-metal",
-  "industrial",
-  "symphonic",
-  "hair-metal",
-  "gothic",
-  "drone",
-  "speed-metal",
-  "power-metal",
-  "psychobilly",
-  "darkwave",
-  "post-punk",
-  "shoegaze",
-  "grindcore",
-  "blackened-jazz",
-  "viking-metal",
-  "pirate-metal",
-  "doom-jazz",
-  "djent",
-  "crust-punk",
-  "dream-pop",
-  "math-rock",
-  "noise-rock",
-  "krautrock",
-  "post-rock",
-  "post-metal",
-  "blackgaze",
-  "deathcore",
-  "metalcore",
-  "groovecore",
-  "nu-metal",
-  "rap-metal",
-  "country-metal",
-  "celtic-metal",
-  "medieval-metal",
-  "baroque-metal",
-  "circus-metal",
-  "chiptune-metal",
-  "synthwave",
-  "electroclash",
-  "horror-punk",
-  "cowpunk",
-  "skiffle-metal",
-  "dungeon-synth",
-  "bedroom-doom",
-  "garage-thrash",
-  "polka-grind",
-  "screamo",
-  "emo-violence",
-  "math-core",
-  "post-hardcore",
-  "djent-folk",
-  "chamber-doom",
-  "zydeco-thrash",
-  "sea-shanty-core",
-] as const;
