--- src/shared/band-name-generator/intensities.ts
@@ -1,38 +1,0 @@
-/** Hyperbolic adjective phrases for sensory descriptions */
-export const INTENSITIES = [
-  "ear-splitting",
-  "brain-melting",
-  "head-splitting",
-  "ground-shaking",
-  "soul-crushing",
-  "deafening",
-  "earth-shattering",
-  "skin-peeling",
-  "bone-rattling",
-  "chest-thumping",
-  "thoroughly inadvisable",
-  "barely-legal",
-  "trouser-flapping",
-  "hair-raising",
-  "spleen-loosening",
-  "neighbour-bothering",
-  "windscreen-shattering",
-  "pavement-cracking",
-  "ceiling-loosening",
-  "double-glazing-defeating",
-  "rib-rattling",
-  "cardigan-blowing",
-  "biscuit-tin-disturbing",
-  "regrettably enthusiastic",
-  "unreasonably cheerful",
-  "borderline-dangerous",
-  "questionably-permitted",
-  "absurdly-loud",
-  "unaccountably-aggressive",
-  "knee-buckling",
-  "teeth-shaking",
-  "moustache-curling",
-  "ozone-depleting",
-  "hedge-trimming",
-  "impressively-illegal",
-] as const;
