--- src/shared/band-name-generator/noise-verbs.ts
@@ -1,35 +1,0 @@
-/** Verbs for "the band ___ed the venue" */
-export const NOISE_VERBS = [
-  "invade",
-  "descend on",
-  "tear through",
-  "assault",
-  "demolish",
-  "conquer",
-  "devastate",
-  "obliterate",
-  "occupy",
-  "set fire to",
-  "lay siege to",
-  "blockade",
-  "annex",
-  "decommission",
-  "rearrange",
-  "renegotiate",
-  "remodel",
-  "reupholster",
-  "redecorate",
-  "fumigate",
-  "rezone",
-  "dismantle",
-  "reassemble",
-  "abscond from",
-  "infiltrate",
-  "depopulate",
-  "haunt",
-  "stalk",
-  "loiter in",
-  "graffiti",
-  "monopolise",
-  "outlaw",
-] as const;
