--- src/shared/band-name-generator/number-words.ts
@@ -1,23 +1,0 @@
-/** Spelled-out small numbers for descriptions */
-export const NUMBER_WORDS = [
-  "two",
-  "three",
-  "four",
-  "five",
-  "six",
-  "seven",
-  "eight",
-  "nine",
-  "ten",
-  "eleven",
-  "twelve",
-  "thirteen",
-  "seventeen",
-  "twenty-one",
-  "thirty",
-  "an alarming number of",
-  "an undisclosed number of",
-  "no fewer than four",
-  "approximately seven",
-  "exactly one",
-] as const;
