--- src/shared/band-name-generator/odd-instruments.ts
@@ -1,43 +1,0 @@
-/** Wildly out-of-place instruments for novelty descriptions */
-export const ODD_INSTRUMENTS = [
-  "theremin",
-  "ukulele",
-  "bagpipes",
-  "hurdy-gurdy",
-  "kazoo",
-  "glockenspiel",
-  "accordion",
-  "harp",
-  "sitar",
-  "mandolin",
-  "stylophone",
-  "didgeridoo",
-  "vacuum cleaner",
-  "Casio keyboard",
-  "harpsichord",
-  "calliope",
-  "musical saw",
-  "spoons",
-  "washboard",
-  "jaw harp",
-  "ocarina",
-  "pan flute",
-  "alphorn",
-  "glass harmonica",
-  "lap steel",
-  "slide whistle",
-  "swanee whistle",
-  "wobble board",
-  "Stylophone Beatbox",
-  "single bagpipe drone",
-  "broken Wurlitzer",
-  "ironing board with strings",
-  "rubber-band orchestra",
-  "tuned biscuit tin",
-  "borrowed cello",
-  "untuned banjo",
-  "kettle and spoon",
-  "vintage ondes Martenot",
-  "bicycle bell choir",
-  "amplified beard",
-] as const;
