--- src/shared/band-name-generator/person-names.ts
@@ -1,107 +1,0 @@
-/** First names used for "{Adjective} {Name} and Friends"-style patterns */
-export const BAND_PERSON_NAMES = [
-  "David",
-  "Ozzy",
-  "Lemmy",
-  "Dio",
-  "Axl",
-  "Slash",
-  "Iggy",
-  "Bowie",
-  "Bruce",
-  "Eddie",
-  "Glenn",
-  "Henry",
-  "Tommy",
-  "Johnny",
-  "Ronnie",
-  "Vincent",
-  "Alice",
-  "Stevie",
-  "Jimmy",
-  "Roger",
-  "Trevor",
-  "Brian",
-  "Geoff",
-  "Maureen",
-  "Pamela",
-  "Cassandra",
-  "Susan",
-  "Beverly",
-  "Gladys",
-  "Doris",
-  "Mildred",
-  "Wilma",
-  "Bernard",
-  "Norman",
-  "Reginald",
-  "Cyril",
-  "Mortimer",
-  "Algernon",
-  "Cornelius",
-  "Bartholomew",
-  "Archibald",
-  "Percival",
-  "Horatio",
-  "Egbert",
-  "Lionel",
-  "Clive",
-  "Roland",
-  "Nigel",
-  "Wilbur",
-  "Cuthbert",
-  "Humphrey",
-  "Cedric",
-  "Stanley",
-  "Walter",
-  "Sidney",
-  "Herbert",
-  "Edmund",
-  "Reuben",
-  "Otto",
-  "Fritz",
-  "Boris",
-  "Igor",
-  "Wolfgang",
-  "Heinrich",
-  "Ingrid",
-  "Astrid",
-  "Brunhilda",
-  "Helga",
-  "Ursula",
-  "Greta",
-  "Magda",
-  "Edith",
-  "Ethel",
-  "Hilda",
-  "Beatrice",
-  "Constance",
-  "Prudence",
-  "Patience",
-  "Tabitha",
-  "Agatha",
-  "Hortense",
-  "Drusilla",
-  "Lavinia",
-  "Octavia",
-  "Esmeralda",
-  "Mavis",
-  "Dolores",
-  "Phyllis",
-  "Vera",
-  "Mabel",
-  "Pearl",
-  "Marjorie",
-  "Gertrude",
-  "Lucille",
-  "Verity",
-  "Cordelia",
-  "Imelda",
-  "Reinhold",
-  "Bjorn",
-  "Sven",
-  "Ulrich",
-  "Klaus",
-  "Hans",
-  "Dieter",
-] as const;
