--- src/shared/band-name-generator/prizes.ts
@@ -1,23 +1,0 @@
-/** Karaoke / dress-up prize categories */
-export const PRIZES = [
-  "biggest hair",
-  "loudest scream",
-  "best devil horns",
-  "most convincing growl",
-  "longest air-guitar solo",
-  "deepest voice",
-  "most denim worn at once",
-  "most authentic mullet",
-  "most enthusiastic windmill arm",
-  "best fake corpse-paint",
-  "loudest unaccompanied howl",
-  "longest sustained note",
-  "most accurate Lemmy impression",
-  "fastest tambourine work",
-  "most dramatic hair flip",
-  "most-improved scream since last time",
-  "best-disguised vocal warm-up",
-  "most committed knee-slide",
-  "most emphatic horns-up gesture",
-  "best Joe-Cocker-style face",
-] as const;
