--- src/shared/band-name-generator/prohibitions.ts
@@ -1,36 +1,0 @@
-/** Things that are absolutely not allowed at this gig */
-export const PROHIBITIONS = [
-  "stage diving onto the pizza",
-  "phones above shoulder height",
-  "vuvuzelas",
-  "glitter cannons",
-  "throwing pints",
-  "uninvited cowbell",
-  "amateur pyrotechnics",
-  "crowdsurfing inflatables",
-  "audience requests",
-  "pretending to know the lyrics",
-  "shouting Wonderwall",
-  "encore-baiting",
-  "bringing your own setlist",
-  "open-toed footwear in the pit",
-  "drum solos by attendees",
-  "tribute-act T-shirts",
-  "loud crisp packets",
-  "moshing during ballads",
-  "talking during ballads",
-  "filming with two phones at once",
-  "selfie sticks of any length",
-  "questioning the headliner's setlist",
-  "improvised lyrics",
-  "queue-jumping at merch",
-  "uninvited stage invasions",
-  "balloon animals",
-  "vapes the size of a kettle",
-  "outdoor flares",
-  "professional cameras without a pass",
-  "harmonicas brought from home",
-  "negotiating refunds mid-set",
-  "audible sighs during solos",
-  "inappropriately enthusiastic kazoo work",
-] as const;
