--- src/shared/band-name-generator/requirements.ts
@@ -1,31 +1,0 @@
-/** Items of clothing or attire that are notionally compulsory */
-export const REQUIREMENTS = [
-  "sequins",
-  "leather",
-  "denim",
-  "mascara",
-  "hairspray",
-  "shoulder pads",
-  "bullet belts",
-  "tasselled jackets",
-  "shredded jeans",
-  "spandex",
-  "corpse paint",
-  "studded gauntlets",
-  "pirate hats",
-  "viking helmets",
-  "novelty cravats",
-  "ill-fitting capes",
-  "platform boots",
-  "bandanas (any colour)",
-  "ironic moustaches",
-  "elaborate eyeliner",
-  "battle vests",
-  "back patches",
-  "fingerless gloves",
-  "wraparound sunglasses",
-  "kohl-rimmed eyes",
-  "dramatic neckwear",
-  "regrettable haircuts",
-  "tour T-shirts (preferably someone else's)",
-] as const;
