--- src/shared/band-name-generator/show-items.ts
@@ -1,54 +1,0 @@
-/** Concrete things you'd see / smell / regret at a gig */
-export const SHOW_ITEMS = [
-  "denim",
-  "leather",
-  "mead",
-  "regret",
-  "glory",
-  "smoke",
-  "lasers",
-  "dry ice",
-  "bandanas",
-  "studs",
-  "shredding solos",
-  "blast beats",
-  "double kicks",
-  "fog machines",
-  "questionable lyrics",
-  "obscure cover versions",
-  "extended drum solos",
-  "merch you'll regret on Monday",
-  "improvised pyrotechnics",
-  "spilled beer",
-  "spandex",
-  "corpse paint",
-  "bullet belts",
-  "leather kilts",
-  "body glitter",
-  "patchouli",
-  "warm lager",
-  "cold sausage rolls",
-  "merchandise queues",
-  "guitar picks",
-  "drumstick souvenirs",
-  "a single rogue bagpipe",
-  "exhausted sound engineers",
-  "questionable hairlines",
-  "ill-advised tattoos",
-  "second-hand smoke",
-  "third-hand stories",
-  "sweat-stained setlists",
-  "audible feedback",
-  "inaudible vocals",
-  "broken cymbals",
-  "duct tape",
-  "split jeans",
-  "torn t-shirts",
-  "missing teeth",
-  "discounted earplugs",
-  "questionable life choices",
-  "everyone's last shred of dignity",
-  "free posters",
-  "expired pizza vouchers",
-  "improvised mosh choreography",
-] as const;
