--- src/shared/band-name-generator/suffixes.ts
@@ -1,88 +1,0 @@
-/** Tail phrases that turn a noun phrase into a fully-realised band name */
-export const BAND_SUFFIXES = [
-  "of Doom",
-  "of Mercy",
-  "of Sorrow",
-  "of the Damned",
-  "of the Forsaken",
-  "and Friends",
-  "Reunited",
-  "Unleashed",
-  "Resurrected",
-  "Returns",
-  "Forever",
-  "Reborn",
-  "Triumphant",
-  "Ascendant",
-  "of the Apocalypse",
-  "of the Void",
-  "of the Storm",
-  "of the Abyss",
-  "of the North",
-  "of Chaos",
-  "of Madness",
-  "of Eternity",
-  "of the Crypt",
-  "and the Disciples",
-  "and the Acolytes",
-  "and the Pilgrims",
-  "and Sons",
-  "and Daughters",
-  "Experience",
-  "Project",
-  "Collective",
-  "Brigade",
-  "Brotherhood",
-  "Sisterhood",
-  "Cult",
-  "Coven",
-  "Convocation",
-  "Conspiracy",
-  "Manifesto",
-  "Revival",
-  "Live in Concert",
-  "MMXXVI",
-  "of the South",
-  "of the East",
-  "of the West",
-  "of Yesterday",
-  "of Tomorrow",
-  "of the Underworld",
-  "of the Overworld",
-  "of the Marshlands",
-  "of the Highlands",
-  "of the Lowlands",
-  "of the Swamp",
-  "of the Steppe",
-  "of the Tundra",
-  "of the Ravine",
-  "Reanimated",
-  "Refurbished",
-  "Recommissioned",
-  "Reconvened",
-  "Re-Released",
-  "and the Quartermasters",
-  "and the Apothecaries",
-  "and the Notaries",
-  "and the Clerks",
-  "and the Custodians",
-  "Quintet",
-  "Sextet",
-  "Septet",
-  "Orchestra",
-  "Ensemble",
-  "Society",
-  "Guild",
-  "Syndicate",
-  "Authority",
-  "Committee",
-  "Tribunal",
-  "Inquest",
-  "Convention",
-  "Roadshow",
-  "Caravan",
-  "Pageant",
-  "Processional",
-  "Wake",
-  "Vigil",
-] as const;
