--- src/shared/band-name-generator/times.ts
@@ -1,13 +1,0 @@
-/** Common gig start times (24h-style hours) */
-export const TIMES = [
-  "6",
-  "7",
-  "7:30",
-  "8",
-  "8:30",
-  "9",
-  "9:30",
-  "10",
-  "11",
-  "midnight",
-] as const;
