--- src/shared/band-name-generator/tour-adjectives.ts
@@ -1,28 +1,0 @@
-/** Adjectives that qualify a "tour" or one-off comeback show */
-export const TOUR_ADJECTIVES = [
-  "farewell",
-  "reunion",
-  "comeback",
-  "retrospective",
-  "anniversary",
-  "definitely-the-last",
-  "victory",
-  "redemption",
-  "vengeance",
-  "valedictory",
-  "absolutely-final",
-  "really-really-last",
-  "we-mean-it-this-time",
-  "first and last",
-  "swansong",
-  "homecoming",
-  "tearful",
-  "long-awaited",
-  "long-promised",
-  "long-overdue",
-  "mostly-original-lineup",
-  "approximately-original-lineup",
-  "one-night-only",
-  "two-nights-only",
-  "open-ended",
-] as const;
