--- src/shared/band-name-generator/venue-types.ts
@@ -1,64 +1,0 @@
-/** Indoor venue archetypes — concrete buildings/rooms */
-export const VENUE_TYPES = [
-  "Arena",
-  "Academy",
-  "Stadium",
-  "Auditorium",
-  "Hall",
-  "Theatre",
-  "Coliseum",
-  "Amphitheatre",
-  "Pavilion",
-  "Forum",
-  "Dome",
-  "Centre",
-  "Roundhouse",
-  "Tabernacle",
-  "Sanctuary",
-  "Crypt",
-  "Vault",
-  "Bunker",
-  "Warehouse",
-  "Factory",
-  "Mill",
-  "Foundry",
-  "Quarry",
-  "Depot",
-  "Garage",
-  "Lounge",
-  "Club",
-  "Den",
-  "Pit",
-  "Chamber",
-  "Ballroom",
-  "Workingmen's Club",
-  "Social Club",
-  "Working Men's Institute",
-  "Bingo Hall",
-  "Snooker Hall",
-  "Boxing Gym",
-  "Scout Hut",
-  "Village Hall",
-  "Lecture Theatre",
-  "Cinema",
-  "Skating Rink",
-  "Roller Disco",
-  "Bowling Alley",
-  "Mausoleum",
-  "Crematorium",
-  "Conservatory",
-  "Pump House",
-  "Boiler Room",
-  "Engine Shed",
-  "Tram Shed",
-  "Toll House",
-  "Custom House",
-  "Drill Hall",
-  "Masonic Lodge",
-  "Reading Room",
-  "Tea Rooms",
-  "Assembly Rooms",
-  "Music Room",
-  "Practice Room",
-  "Rehearsal Studio",
-] as const;
