--- src/shared/band-name-generator/weird-venues.ts
@@ -1,42 +1,0 @@
-/** Improbable buildings to be "converted into" a venue */
-export const WEIRD_VENUES = [
-  "lift shaft",
-  "granary",
-  "bus depot",
-  "brewery vat",
-  "postal sorting office",
-  "cattle market",
-  "abandoned multi-storey",
-  "decommissioned chapel",
-  "1980s leisure centre",
-  "shipping container",
-  "village butcher's freezer",
-  "art-deco cinema",
-  "Victorian bathhouse",
-  "Edwardian bandstand",
-  "disused fire station",
-  "former post office",
-  "redundant church",
-  "out-of-service ferry",
-  "decommissioned lighthouse",
-  "abandoned underground station",
-  "WW2 pillbox",
-  "converted barn",
-  "haunted greenhouse",
-  "old reading room",
-  "empty Woolworths",
-  "former Blockbuster",
-  "former HMV",
-  "shut-down bowling alley",
-  "padlocked snooker hall",
-  "boarded-up pub",
-  "derelict library",
-  "scrapped trawler",
-  "retired narrowboat",
-  "carriage-works",
-  "pump house",
-  "abandoned railway tunnel",
-  "long-suffering scout hut",
-  "regrettable nightclub",
-  "decommissioned petrol station",
-] as const;
